pyproject.toml
[project]
name = "ml-waf-anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "streamlit>=1.52.2",
]
